+++ R/sim_out.R
@@ -168,14 +168,14 @@ print_extra <- function(x) {
   cat(header_str("Element(s) of the First Dataset",
                  hw = .5,
                  prefix = "\n",
-                 suffix = "\n"))
+                 suffix = "\n\n"))
 
   for (xx in extra_names) {
     tmp <- paste0("<", xx, ">")
     cat(header_str(tmp,
                    hw = .4,
-                   prefix = "\n",
-                   suffix = "\n\n"))
+                   prefix = "",
+                   suffix = "\n"))
     print(extra[[xx]])
   }
   return(NULL)