+++ NEWS.md
@@ -123,3 +123,7 @@
 
 - Added a print method for `power4test`
   objects. (0.0.0.9027)
+
+- `sim_out` can now be used to update
+  a `sim_out` object by adding elements
+  to `extra`. (0.0.0.9028)